+++ src/aiod/calls/urls.py
@@ -2,7 +2,7 @@ import urllib
 
 from typing import Literal
 
-from aiod_sdk.config.settings import api_base_url, latest_version
+from aiod.config.settings import api_base_url, latest_version
 
 
 def url_to_get_asset(
