+++ src/aiod/resources/computational_assets.py
@@ -1,4 +1,4 @@
-from aiod_sdk.calls import calls
+from aiod.calls import calls
 
 (
     get_list,
@@ -8,4 +8,4 @@ from aiod_sdk.calls import calls
     get_content,
     get_assets_async,
     get_list_async,
-) = calls.wrap_common_calls(asset_type="case_studies")
+) = calls.wrap_common_calls(asset_type="computational_assets")
